move posts from componenets to containers
