+++ src/containers/Posts/Posts.js
@@ -9,7 +9,7 @@ import {
     Button,
     TextField,
 } from '@material-ui/core';
-import Post from './Post/Post';
+import Post from '../../components/Posts/Post/Post';
 
 class Posts extends Component {
     constructor(props) {
